Refactor component structure and put css files in separate directories
