+++ src/components/Footer.js
@@ -1,4 +1,4 @@
-import "./Footer.css";
+import "./styles/Footer.css";
 
 export default function Footer() {
   return (
