+++ src/components/styles/SearchBar.css
