+++ src/pages/styles/FolderPage.css
@@ -1,0 +1,13 @@
+.folder-page-contents {
+  max-width: 1124px;
+  padding: 0 32px;
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  margin: 40px 0;
+  @media (max-width: 784px) {
+    margin: 20px 0;
+  }
+}
