+++ src/pages/styles/SharedPage.css
